backend/app/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
psycopg2-binary==2.8.5
dj-database-url==0.5.0
gunicorn==20.0.4